min function: given some parameters, returns the smallest parameter.
in real use cases, it only makes sense to have a mix of fixed and relative size units as parameters: in the case of 100% 300px, it means that while 100% is less than 300px, it will take up 100% in width, so all the space available. but the moment that exceeds 300px, it sticks with that size.