you need to use transform translate -50% -50% to center it.
translate is a function that moves an element by the x and y parameters it has been given.

We need this, because by default, setting the top and left values each to 50%, moves it a little over halfways, and instead of the middle of the text being in the center, the top and left edges are. When moving it back by -50 -50, we move it by half the elements width.